Fixed ProfileScreen validation, fixed OnBoarding validation, minor fixes to onboarding and profile UI

+++ app/src/main/java/it/unito/progmob/home/presentation/components/CircularProgressBar.kt
@@ -51,7 +51,7 @@ fun CircularProgressBar(
     }
 
     val stepCircularBarAnimation = animateFloatAsState(
-        targetValue = if (stepBarAnimationPlayed) (((steps * 100) / targetStepsGoal) / 100).toFloat() else 0f,
+        targetValue = if (stepBarAnimationPlayed) ((steps * 100) / targetStepsGoal) / 100f else 0f,
         animationSpec = tween(
             durationMillis = animDuration,
             delayMillis = animDelay
